requirements.txt
matplotlib==3.8.4
networkx==3.2.1
numpy==2.3.2
python_igraph==0.11.8
scikit_learn==1.4.2
scipy==1.16.1
torch==2.5.1
torch_geometric==2.6.1
torch_scatter==2.1.2+pt25cu124
